Split apolloUnmask tests
