+++ packages/plugins/typescript/operations/tests/ts-documents.spec.ts
@@ -3,92 +3,11 @@ import { validateTs } from '@graphql-codegen/testing';
 import { buildClientSchema, buildSchema, parse } from 'graphql';
 import { plugin as tsPlugin } from '../../typescript/src/index.js';
 import { plugin } from '../src/index.js';
+import { schema } from './shared/schema.js';
 
 describe('TypeScript Operations Plugin', () => {
   // eslint-disable-next-line @typescript-eslint/no-require-imports
   const gitHuntSchema = buildClientSchema(require('../../../../../dev-test/githunt/schema.json'));
-
-  const schema = buildSchema(/* GraphQL */ `
-    scalar DateTime
-
-    input InputType {
-      t: String
-    }
-
-    type User {
-      id: ID!
-      username: String!
-      email: String!
-      profile: Profile
-      role: Role
-    }
-
-    type Profile {
-      age: Int
-      firstName: String!
-    }
-
-    type Mutation {
-      test: String
-      login(username: String!, password: String!): User
-    }
-
-    type Subscription {
-      userCreated: User
-    }
-
-    interface Notifiction {
-      id: ID!
-      createdAt: String!
-    }
-
-    type TextNotification implements Notifiction {
-      id: ID!
-      text: String!
-      createdAt: String!
-    }
-
-    type ImageNotification implements Notifiction {
-      id: ID!
-      imageUrl: String!
-      metadata: ImageMetadata!
-      createdAt: String!
-    }
-
-    type ImageMetadata {
-      createdBy: String!
-    }
-
-    enum Role {
-      USER
-      ADMIN
-    }
-
-    union MyUnion = User | Profile
-
-    union AnyNotification = TextNotification | ImageNotification
-    union SearchResult = TextNotification | ImageNotification | User
-
-    type Query {
-      me: User
-      unionTest: MyUnion
-      notifications: [Notifiction!]!
-      mixedNotifications: [AnyNotification!]!
-      search(term: String!): [SearchResult!]!
-      dummy: String
-      dummyNonNull: String!
-      dummyArray: [String]
-      dummyNonNullArray: [String]!
-      dummyNonNullArrayWithValues: [String!]!
-      dummyWithType: Profile
-    }
-
-    schema {
-      query: Query
-      mutation: Mutation
-      subscription: Subscription
-    }
-  `);
 
   const validate = async (
     content: Types.PluginOutput,
@@ -7422,169 +7341,6 @@ function test(q: GetEntityBrandDataQuery): void {
           ) | null };
 
         export type UserFragmentFragment = { __typename?: 'User', id: string } & { ' $fragmentName'?: 'UserFragmentFragment' };
-      `);
-    });
-
-    it("'mask' with @unmask configured with apolloUnmask yields correct types", async () => {
-      const ast = parse(/* GraphQL */ `
-        query {
-          me {
-            ...UserFragment @unmask
-          }
-        }
-        fragment UserFragment on User {
-          id
-        }
-      `);
-      const result = await plugin(
-        schema,
-        [{ location: 'test-file.ts', document: ast }],
-        { inlineFragmentTypes: 'mask', customDirectives: { apolloUnmask: true } },
-        { outputFile: '' }
-      );
-      expect(result.content).toBeSimilarStringTo(`
-        export type Unnamed_1_QueryVariables = Exact<{ [key: string]: never; }>;
-
-
-        export type Unnamed_1_Query = { __typename?: 'Query', me?: (
-            { __typename?: 'User', id: string }
-            & { ' $fragmentRefs'?: { 'UserFragmentFragment': UserFragmentFragment } }
-          ) | null };
-
-        export type UserFragmentFragment = { __typename?: 'User', id: string } & { ' $fragmentName'?: 'UserFragmentFragment' };
-      `);
-    });
-
-    it("'mask' with @unmask without apolloUnmask yields correct types", async () => {
-      const ast = parse(/* GraphQL */ `
-        query {
-          me {
-            ...UserFragment @unmask
-          }
-        }
-        fragment UserFragment on User {
-          id
-        }
-      `);
-      const result = await plugin(
-        schema,
-        [{ location: 'test-file.ts', document: ast }],
-        { inlineFragmentTypes: 'mask' },
-        { outputFile: '' }
-      );
-      expect(result.content).toBeSimilarStringTo(`
-        export type Unnamed_1_QueryVariables = Exact<{ [key: string]: never; }>;
-
-
-        export type Unnamed_1_Query = { __typename?: 'Query', me?: (
-            { __typename?: 'User' }
-            & { ' $fragmentRefs'?: { 'UserFragmentFragment': UserFragmentFragment } }
-          ) | null };
-
-        export type UserFragmentFragment = { __typename?: 'User', id: string } & { ' $fragmentName'?: 'UserFragmentFragment' };
-      `);
-    });
-
-    it("'mask' with @unmask with apolloUnmask explicitly disabled yields correct types", async () => {
-      const ast = parse(/* GraphQL */ `
-        query {
-          me {
-            ...UserFragment @unmask
-          }
-        }
-        fragment UserFragment on User {
-          id
-        }
-      `);
-      const result = await plugin(
-        schema,
-        [{ location: 'test-file.ts', document: ast }],
-        { inlineFragmentTypes: 'mask', customDirectives: { apolloUnmask: false } },
-        { outputFile: '' }
-      );
-      expect(result.content).toBeSimilarStringTo(`
-        export type Unnamed_1_QueryVariables = Exact<{ [key: string]: never; }>;
-
-
-        export type Unnamed_1_Query = { __typename?: 'Query', me?: (
-            { __typename?: 'User' }
-            & { ' $fragmentRefs'?: { 'UserFragmentFragment': UserFragmentFragment } }
-          ) | null };
-
-        export type UserFragmentFragment = { __typename?: 'User', id: string } & { ' $fragmentName'?: 'UserFragmentFragment' };
-      `);
-    });
-
-    it("'mask' with @unmask and masked fragments yields correct types", async () => {
-      const ast = parse(/* GraphQL */ `
-        query {
-          me {
-            ...UserFragment @unmask
-            ...UserFragment2
-          }
-        }
-        fragment UserFragment on User {
-          id
-        }
-
-        fragment UserFragment2 on User {
-          email
-        }
-      `);
-      const result = await plugin(
-        schema,
-        [{ location: 'test-file.ts', document: ast }],
-        { inlineFragmentTypes: 'mask', customDirectives: { apolloUnmask: true } },
-        { outputFile: '' }
-      );
-      expect(result.content).toBeSimilarStringTo(`
-        export type Unnamed_1_QueryVariables = Exact<{ [key: string]: never; }>;
-
-
-        export type Unnamed_1_Query = { __typename?: 'Query', me?: (
-            { __typename?: 'User', id: string }
-            & { ' $fragmentRefs'?: { 'UserFragmentFragment': UserFragmentFragment;'UserFragment2Fragment': UserFragment2Fragment } }
-          ) | null };
-
-        export type UserFragmentFragment = { __typename?: 'User', id: string } & { ' $fragmentName'?: 'UserFragmentFragment' };
-        export type UserFragment2Fragment = { __typename?: 'User', email: string } & { ' $fragmentName'?: 'UserFragment2Fragment' };
-      `);
-    });
-
-    it("'mask' with @unmask and masked fragments on overlapping fields yields correct types", async () => {
-      const ast = parse(/* GraphQL */ `
-        query {
-          me {
-            ...UserFragment @unmask
-            ...UserFragment2
-          }
-        }
-        fragment UserFragment on User {
-          id
-          email
-        }
-
-        fragment UserFragment2 on User {
-          email
-        }
-      `);
-      const result = await plugin(
-        schema,
-        [{ location: 'test-file.ts', document: ast }],
-        { inlineFragmentTypes: 'mask', customDirectives: { apolloUnmask: true } },
-        { outputFile: '' }
-      );
-      expect(result.content).toBeSimilarStringTo(`
-        export type Unnamed_1_QueryVariables = Exact<{ [key: string]: never; }>;
-
-
-        export type Unnamed_1_Query = { __typename?: 'Query', me?: (
-            { __typename?: 'User', id: string, email: string }
-            & { ' $fragmentRefs'?: { 'UserFragmentFragment': UserFragmentFragment;'UserFragment2Fragment': UserFragment2Fragment } }
-          ) | null };
-
-        export type UserFragmentFragment = { __typename?: 'User', id: string, email: string } & { ' $fragmentName'?: 'UserFragmentFragment' };
-        export type UserFragment2Fragment = { __typename?: 'User', email: string } & { ' $fragmentName'?: 'UserFragment2Fragment' };
       `);
     });
   });
